# app/backend/app/routers/candidates.py
from fastapi import APIRouter, Depends, HTTPException, Query, status
from sqlmodel.ext.asyncio.session import AsyncSession

from app.database import get_session
from app.dependencies.auth import get_current_user
from app.exceptions import ConflictError, NotFoundException, ValidationError
from app.models.user import User
from app.schemas.candidates import (
    CandidateCreate,
    CandidateDetailResponse,
    CandidateListItem,
    CandidatePositionCreate,
    CandidatePositionResponse,
    CandidateResponse,
    CandidateUpdate,
    PaginatedCandidates,
    PositionStageItem,
    StageUpdate,
)
from app.schemas.documents import DocumentResponse
from app.services import candidate_service, document_service

router = APIRouter(prefix="/api/candidates", tags=["candidates"])


@router.get("", response_model=PaginatedCandidates)
async def list_candidates(
    offset: int = Query(default=0, ge=0),
    limit: int = Query(default=20, ge=1, le=100),
    search: str | None = Query(default=None),
    stage: str | None = Query(default=None),
    position_id: int | None = Query(default=None),
    sort_by: str | None = Query(default=None),
    sort_order: str | None = Query(default=None),
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> PaginatedCandidates:
    items, total = await candidate_service.list_candidates(
        session, offset, limit, search, stage, position_id, sort_by, sort_order
    )
    return PaginatedCandidates(
        items=[
            CandidateListItem(
                id=item["id"],
                full_name=item["full_name"],
                email=item["email"],
                positions=[PositionStageItem(**pos) for pos in item["positions"]],
                updated_at=item["updated_at"],
            )
            for item in items
        ],
        total=total,
        offset=offset,
        limit=limit,
    )


@router.post("", response_model=CandidateResponse, status_code=status.HTTP_201_CREATED)
async def create_candidate(
    body: CandidateCreate,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidateResponse:
    try:
        candidate = await candidate_service.create_candidate(
            session,
            body.full_name,
            body.email,
        )
        return CandidateResponse(
            id=candidate.id,
            full_name=candidate.full_name,
            email=candidate.email,
            is_archived=candidate.is_archived,
        )
    except ConflictError as e:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail=e.detail,
        ) from e


@router.get("/{candidate_id}", response_model=CandidateDetailResponse)
async def get_candidate(
    candidate_id: int,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidateDetailResponse:
    detail = await candidate_service.get_candidate_detail(session, candidate_id)
    if detail is None:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Candidate not found",
        )

    positions = [PositionStageItem(**pos) for pos in detail["positions"]]

    return CandidateDetailResponse(
        id=detail["id"],
        full_name=detail["full_name"],
        email=detail["email"],
        is_archived=detail["is_archived"],
        positions=positions,
        created_at=detail["created_at"],
        updated_at=detail["updated_at"],
    )


@router.patch("/{candidate_id}", response_model=CandidateResponse)
async def update_candidate(
    candidate_id: int,
    body: CandidateUpdate,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidateResponse:
    try:
        candidate = await candidate_service.update_candidate(
            session,
            candidate_id,
            body.full_name,
            body.email,
        )
        return CandidateResponse(
            id=candidate.id,
            full_name=candidate.full_name,
            email=candidate.email,
            is_archived=candidate.is_archived,
        )
    except NotFoundException as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=e.detail,
        ) from e
    except ConflictError as e:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail=e.detail,
        ) from e


@router.post(
    "/{candidate_id}/positions",
    response_model=CandidatePositionResponse,
    status_code=status.HTTP_201_CREATED,
)
async def add_to_position(
    candidate_id: int,
    body: CandidatePositionCreate,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidatePositionResponse:
    try:
        candidate_position = await candidate_service.add_to_position(
            session,
            candidate_id,
            body.position_id,
        )
        return CandidatePositionResponse(
            id=candidate_position.id,
            candidate_id=candidate_position.candidate_id,
            position_id=candidate_position.position_id,
            stage=candidate_position.stage,
            created_at=candidate_position.created_at.isoformat(),
            updated_at=candidate_position.updated_at.isoformat(),
        )
    except NotFoundException as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=e.detail,
        ) from e
    except ConflictError as e:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail=e.detail,
        ) from e


@router.delete(
    "/{candidate_id}/positions/{position_id}", status_code=status.HTTP_204_NO_CONTENT
)
async def remove_from_position(
    candidate_id: int,
    position_id: int,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> None:
    try:
        await candidate_service.remove_from_position(
            session,
            candidate_id,
            position_id,
        )
    except NotFoundException as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=e.detail,
        ) from e


@router.patch(
    "/{candidate_id}/positions/{position_id}", response_model=CandidatePositionResponse
)
async def update_stage(
    candidate_id: int,
    position_id: int,
    body: StageUpdate,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidatePositionResponse:
    """Update pipeline stage. Return 404 if not found, 422 if invalid transition."""
    try:
        candidate_position = await candidate_service.update_stage(
            session,
            candidate_id,
            position_id,
            body.stage,
        )
        return CandidatePositionResponse(
            id=candidate_position.id,
            candidate_id=candidate_position.candidate_id,
            position_id=candidate_position.position_id,
            stage=candidate_position.stage,
            created_at=candidate_position.created_at.isoformat(),
            updated_at=candidate_position.updated_at.isoformat(),
        )
    except NotFoundException as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=e.detail,
        ) from e
    except ValidationError as e:
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
            detail=e.detail,
        ) from e


@router.post("/{candidate_id}/archive", response_model=CandidateResponse)
async def archive_candidate(
    candidate_id: int,
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> CandidateResponse:
    try:
        candidate = await candidate_service.archive_candidate(session, candidate_id)
        return CandidateResponse(
            id=candidate.id,
            full_name=candidate.full_name,
            email=candidate.email,
            is_archived=candidate.is_archived,
        )
    except NotFoundException as e:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail=e.detail,
        ) from e


@router.get("/{candidate_id}/documents", response_model=list[DocumentResponse])
async def list_candidate_documents(
    candidate_id: int,
    position_id: int | None = Query(default=None),
    type: str | None = Query(default=None),
    candidate_position_id: int | None = Query(default=None),
    session: AsyncSession = Depends(get_session),
    current_user: User = Depends(get_current_user),
) -> list[DocumentResponse]:
    documents = await document_service.list_candidate_documents(
        session,
        candidate_id,
        position_id,
        type,
        candidate_position_id,
    )
    return [DocumentResponse(**doc) for doc in documents]